Moved code run on worker thread from sync manager to new Runnable subclasses * SyncRunnable (super class) * SyncUpRunnable (subclass of SyncRunnable - runnable for sync up) * AdvancedSyncUpRunnable (subclass of SyncUpRunnable - runnable for advanced sync up) * SyncDownRunnable (subclass of SyncRunnable - runnable for sync down) * CleanSyncGhostsRunnable (subclass of SyncRunnable - runnable for clean sync ghosts)
